requirements.txt
openai
python-dotenv
whisper